pyproject.toml
[project]
name = "farmgamelegal-1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.7",
]
